# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.5"
  title: SiB Colombia - Dataportal API
  description: API for SiB Colombia Dataportal. This API is used as a backend for SiB Colombia's web sites and online services.
  termsOfService: http://swagger.io/terms/
  contact:
    name: SiB Colombia Developer Team
    url: http://www.sibcolombia.net
    email: lgrajales@humboldt.org.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# during dev, should point to your local machine
host: localhost:5000
...
basePath: /api/v1.5
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /occurrence/count:
    # binds a127 app logic to a route
    x-swagger-router-controller: occurrences
    get:
      description: Returns count of occurrences according to query parameters
      tags:
        - occurrence
      # used as the method name of the controller
      operationId: occurrenceCount
      parameters:
        - name: isGeoreferenced
          in: query
          description: True or false if you wants only georeferenced occurrences
          required: false
          type: boolean
        - name: group
          in: query
          description: The access groups to query.
          required: false
          type: string
          default: guess
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResultResponse"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /occurrence/grid:
    x-swagger-router-controller: occurrences
    get:
      description: Returns a grid of cells with occurrences count of each cell according to query parameters
      tags:
        - occurrence
      operationId: gridSearch
      parameters:
        - name: q
          in: query
          description: Simple search parameter. The value for this parameter can be a simple word or a phrase. See http://lucene.apache.org/core/6_1_0/queryparser/org/apache/lucene/queryparser/classic/package-summary.html#package_description
          required: false
          type: string
        - name: group
          in: query
          description: The access groups to query.
          required: false
          type: string
          default: guess
        - name: scientificName
          in: query
          description: Scientific name.
          required: false
          type: array
          items:
            type: string
        - name: kingdomName
          in: query
          description: Kingdom name.
          required: false
          type: array
          items:
            type: string
        - name: phylumName
          in: query
          description: Phylum name.
          required: false
          type: array
          items:
            type: string
        - name: className
          in: query
          description: Class name.
          required: false
          type: array
          items:
            type: string
        - name: orderName
          in: query
          description: Order name.
          required: false
          type: array
          items:
            type: string
        - name: familyName
          in: query
          description: Family name.
          required: false
          type: array
          items:
            type: string
        - name: genusName
          in: query
          description: Genus name.
          required: false
          type: array
          items:
            type: string
        - name: speciesName
          in: query
          description: Species name.
          required: false
          type: array
          items:
            type: string
        - name: specificEpithetName
          in: query
          description: Specific epithet name.
          required: false
          type: array
          items:
            type: string
        - name: infraspecificEpithetName
          in: query
          description: Infraspecific epithet name.
          required: false
          type: array
          items:
            type: string
        - name: providerName
          in: query
          description: Provider name.
          required: false
          type: array
          items:
            type: string
        - name: resourceName
          in: query
          description: Resource name.
          required: false
          type: array
          items:
            type: string
        - name: collectionName
          in: query
          description: Collection name.
          required: false
          type: array
          items:
            type: string
        - name: institutionCode
          in: query
          description: Institution code.
          required: false
          type: array
          items:
            type: string
        - name: countryName
          in: query
          description: Country name.
          required: false
          type: array
          items:
            type: string
        - name: departmentName
          in: query
          description: Department name.
          required: false
          type: array
          items:
            type: string
        - name: countyName
          in: query
          description: County name.
          required: false
          type: array
          items:
            type: string
        - name: habitat
          in: query
          description: Habitat.
          required: false
          type: array
          items:
            type: string
        - name: basisOfRecord
          in: query
          description: Basis of record.
          required: false
          type: array
          items:
            type: string
        - name: elevationEqual
          in: query
          description: "Occurrence elevation. This query parameter searchs for elevation equal to a number."
          required: false
          type: array
          items:
            type: number
            format: double
        - name: elevationGreaterOrEqualTo
          in: query
          description: "Occurrence elevation. This query parameter searchs for elevation equal or greater than a number."
          type: array
          items:
            type: number
            format: double
          required: false
        - name: elevationLessOrEqualTo
          in: query
          description: "Occurrence elevation. This query parameter searchs for elevation equal or less than a number."
          type: array
          items:
            type: number
            format: double
          required: false
        - name: depthEqual
          in: query
          description: "Occurrence depth. This query parameter searchs for depth equal to a number."
          type: array
          items:
            type: number
            format: double
          required: false
        - name: depthGreaterOrEqualTo
          in: query
          description: "Occurrence depth. This query parameter searchs for depth equal or greater than a number."
          type: array
          items:
            type: number
            format: double
          required: false
        - name: depthLessOrEqualTo
          in: query
          description: "Occurrence depth. This query parameter searchs for depth equal or less than a number."
          type: array
          items:
            type: number
            format: double
          required: false
        - name: latitudeTopLeft
          in: query
          description: "Data located around square defined by top left point and bottom right point. Use latitudeTopLeft for the latitude of the top left point."
          type: number
          format: double
          required: false
        - name: longitudeTopLeft
          in: query
          description: "Data located around square defined by top left point and bottom right point. Use latitudeTopLeft for the longitude of the top left point."
          type: number
          format: double
          required: false
        - name: latitudeBottomRight
          in: query
          description: "Data located around square defined by top left point and bottom right point. Use latitudeBottomRight for the latitude of the bottom right point."
          type: number
          format: double
          required: false
        - name: longitudeBottomRight
          in: query
          description: "Data located around square defined by top left point and bottom right point. Use longitudeBottomRight for the longitude of the bottom right point."
          type: number
          format: double
          required: false
        - name: precision
          in: query
          description: "Grid hash precision. Valid values. 1: 5,009.4km x 4,992.6km, 2: 1,252.3km x 624.1km, 3: 156.5km x 156km, 4: 39.1km x 19.5km, 5: 4.9km x 4.9km, 6: 1.2km x 609.4m"
          required: true
          type: integer
          default: 5
          minimum: 1
          maximum: 6
        - name: responseType
          in: query
          description: "Type of response for this query, valid value: geojson. Default value geojson"
          required: true
          type: string
          default: "geojson"
        - name: scale
          in: query
          description: "Type of scale for cell coloring, valid values: linear or logarithmic"
          required: true
          type: string
          default: "logarithmic"
        - name: color
          in: query
          description: "Color to apply for each cell."
          required: true
          type: string
          default: "#ff2600"
        - name: colorMethod
          in: query
          description: "Method applied for cell density color, valid values: alpha or gradient"
          required: true
          type: string
          default: "gradient"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/GeoJSON"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  GeoJSON:
    required:
      - type
    properties:
      type:
        type: string
      features:
        type: array
        items:
          $ref: '#/definitions/Feature'
  Feature:
    required:
      - type
    properties:
      type:
        type: string
      geometry:
        $ref: '#/definitions/Geometry'
  Geometry:
    properties:
      type:
        type: string
      coordinates:
        type: array
        items:
          $ref: '#/definitions/Coordinate'
      properties:
        $ref: '#/definitions/GeometryProperties'
  Coordinate:
    type: array
    items:
      type: number
      format: double
  GeometryProperties:
    properties:
      stroke:
        type: string
      stroke-width:
        type: number
        format: float
        minimum: 0
      stroke-opacity:
        type: number
        format: float
        minimum: 0
        maximum: 1
      fill:
        type: string
      fill-opacity:
        type: number
        format: float
        minimum: 0
        maximum: 1
      geohash:
        type: string
      count:
        type: integer
...
  ResultResponse:
    required:
      - count
    properties:
      count:
        type: integer
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
      description:
        type: string